+++ packages/snap-controller/src/Search/SearchController.ts
@@ -6,10 +6,10 @@ import { StorageStore, ErrorType } from '@searchspring/snap-store-mobx';
 import { getSearchParams } from '../utils/getParams';
 import { ControllerTypes, PageContextVariable } from '../types';
 
-import type { Product, Banner, SearchStore } from '@searchspring/snap-store-mobx';
+import type { Product, Banner, SearchStore, ValueFacet } from '@searchspring/snap-store-mobx';
 import type {
 	SearchControllerConfig,
-	AfterSearchObj,
+	SearchAfterSearchObj,
 	AfterStoreObj,
 	ControllerServices,
 	ContextVariables,
@@ -18,16 +18,21 @@ import type {
 	BeforeSearchObj,
 } from '../types';
 import type { Next } from '@searchspring/snap-event-manager';
-import type {
-	SearchRequestModel,
-	SearchResponseModelResult,
-	SearchRequestModelSearchRedirectResponseEnum,
-	MetaResponseModel,
-	SearchResponseModel,
-	SearchRequestModelFilterRange,
-	SearchRequestModelFilterValue,
-	SearchRequestModelFilter,
+import {
+	type SearchRequestModel,
+	type SearchResponseModelResult,
+	type SearchRequestModelSearchRedirectResponseEnum,
+	type MetaResponseModel,
+	type SearchResponseModel,
+	type SearchRequestModelFilterRange,
+	type SearchRequestModelFilterValue,
+	type SearchRequestModelFilter,
+	type SearchResponseModelFilter,
+	type MetaResponseModelFacetHierarchy,
+	type SearchResponseModelFilterTypeEnum,
+	SearchResponseModelFacetValue,
 } from '@searchspring/snapi-types';
+
 import {
 	type AutocompleteAddtocartSchemaDataBgfilterInner,
 	type AutocompleteAddtocartSchemaDataFilterInner,
@@ -41,7 +46,7 @@ import {
 } from '@searchspring/beacon';
 import { CLICK_DUPLICATION_TIMEOUT, isClickWithinProductLink } from '../utils/isClickWithinProductLink';
 
-const BACKGROUND_FILTER_FIELD_MATCHES = ['collection', 'category', 'categories', 'hierarchy'];
+const BACKGROUND_FILTER_FIELD_MATCHES = ['collection', 'category', 'categories', 'hierarchy', 'brand', 'manufacturer'];
 const BACKGROUND_FILTERS_VALUE_FLAGS = [1, 0, '1', '0', 'true', 'false', true, false];
 
 const defaultConfig: SearchControllerConfig = {
@@ -158,7 +163,7 @@ export class SearchController extends AbstractController {
 		});
 
 		// add 'afterSearch' middleware
-		this.eventManager.on('afterSearch', async (search: AfterSearchObj, next: Next): Promise<void | boolean> => {
+		this.eventManager.on('afterSearch', async (search: SearchAfterSearchObj, next: Next): Promise<void | boolean> => {
 			const config = search.controller.config as SearchControllerConfig;
 			const redirectURL = search.response?.merchandising?.redirect;
 			const searchStore = search.controller.store as SearchStore;
@@ -190,6 +195,51 @@ export class SearchController extends AbstractController {
 			this.eventManager.fire('restorePosition', { controller: this, element: elementPosition });
 		});
 
+		const hierarchySettings = this.config.settings?.filters?.hierarchy;
+		if (hierarchySettings && hierarchySettings.enabled) {
+			this.eventManager.on('afterSearch', async (search: SearchAfterSearchObj, next: Next): Promise<void | boolean> => {
+				await next();
+
+				const displayDelimiter = hierarchySettings.displayDelimiter ?? ' / '; // choose delimiter for label
+				const showFullPath = hierarchySettings.showFullPath ?? false; // display full hierarchy path or just the current level
+
+				// add hierarchy filter to filter summary
+				const facets = search.response.facets;
+				if (facets) {
+					facets.forEach((facet) => {
+						if (search.response.meta?.facets && facet.field) {
+							const metaFacet = search.response.meta.facets[facet.field];
+							const dataDelimiter = (metaFacet as MetaResponseModelFacetHierarchy)?.hierarchyDelimiter || ' / ';
+
+							if (metaFacet && metaFacet.display === 'hierarchy' && facet.filtered && (facet as ValueFacet).values?.length > 0) {
+								const filteredValues = (facet as SearchResponseModelFacetValue).values?.filter((val) => val?.filtered === true);
+
+								if (filteredValues && filteredValues.length) {
+									const filterToAdd: SearchResponseModelFilter = {
+										field: facet.field,
+										//escape special charactors used in regex
+										label: showFullPath
+											? (filteredValues[0].value ?? filteredValues[0].label ?? '').replace(
+													new RegExp(dataDelimiter.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'), 'g'),
+													displayDelimiter
+											  )
+											: filteredValues[0].label,
+										type: 'value' as SearchResponseModelFilterTypeEnum.Value,
+									};
+
+									if (search.response.filters) {
+										search.response.filters.push(filterToAdd);
+									} else {
+										search.response.filters = [filterToAdd];
+									}
+								}
+							}
+						}
+					});
+				}
+			});
+		}
+
 		this.eventManager.on('afterStore', async (search: AfterStoreObj, next: Next): Promise<void | boolean> => {
 			await next();
 			const controller = search.controller as SearchController;
@@ -201,14 +251,13 @@ export class SearchController extends AbstractController {
 				if (products.length === 0 && !search.response._cached) {
 					// handle no results
 					const data = getSearchSchemaData({ params: search.request, response: search.response });
+					this.eventManager.fire('track.product.render', { controller: this, trackEvent: data });
 					this.tracker.events[this.page.type].render({ data, siteId: this.config.globals?.siteId });
 				}
 
 				products.forEach((result: Product) => {
 					if (!search.response._cached) {
-						const data = schemaMap[result.id];
-						this.tracker.events[this.page.type].render({ data, siteId: this.config.globals?.siteId });
-						this.eventManager.fire('track.product.render', { controller: this, product: result, trackEvent: data });
+						this.track.product.render(result);
 					}
 					this.events.product[result.id] = this.events.product[result.id] || {};
 					this.events.product[result.id].render = true;
@@ -350,10 +399,10 @@ export class SearchController extends AbstractController {
 				// store position data or empty object
 				this.storage.set('scrollMap', scrollMap);
 				const data = schemaMap[result.id];
+				this.eventManager.fire('track.product.clickThrough', { controller: this, event: e, product: result, trackEvent: data });
 				this.tracker.events[this.page.type].clickThrough({ data, siteId: this.config.globals?.siteId });
 				this.events.product[result.id] = this.events.product[result.id] || {};
 				this.events.product[result.id].clickThrough = true;
-				this.eventManager.fire('track.product.clickThrough', { controller: this, event: e, product: result, trackEvent: data });
 			},
 			click: (e: MouseEvent, result): void => {
 				if (this.events.product[result.id]?.click) {
@@ -378,10 +427,10 @@ export class SearchController extends AbstractController {
 				}
 
 				const data = schemaMap[result.id];
+				this.eventManager.fire('track.product.render', { controller: this, product: result, trackEvent: data });
 				this.tracker.events[this.page.type].render({ data, siteId: this.config.globals?.siteId });
 				this.events.product[result.id] = this.events.product[result.id] || {};
 				this.events.product[result.id].render = true;
-				this.eventManager.fire('track.product.render', { controller: this, product: result, trackEvent: data });
 			},
 			impression: (result: Product): void => {
 				if (this.events.product[result.id]?.impression || !this.events.product[result.id]?.render) {
@@ -389,24 +438,24 @@ export class SearchController extends AbstractController {
 				}
 
 				const data = schemaMap[result.id];
+				this.eventManager.fire('track.product.impression', { controller: this, product: result, trackEvent: data });
 				this.tracker.events[this.page.type].impression({ data, siteId: this.config.globals?.siteId });
 				this.events.product[result.id] = this.events.product[result.id] || {};
 				this.events.product[result.id].impression = true;
-				this.eventManager.fire('track.product.impression', { controller: this, product: result, trackEvent: data });
 			},
 			addToCart: (result: Product): void => {
 				const data = getSearchAddtocartSchemaData({ searchSchemaData: schemaMap[result.id], results: [result] });
+				this.eventManager.fire('track.product.addToCart', { controller: this, product: result, trackEvent: data });
 				this.tracker.events[this.page.type].addToCart({
 					data,
 					siteId: this.config.globals?.siteId,
 				});
-				this.eventManager.fire('track.product.addToCart', { controller: this, product: result, trackEvent: data });
 			},
 		},
 		redirect: (redirectURL: string): void => {
 			const data = getSearchRedirectSchemaData({ redirectURL });
-			this.tracker.events.search.redirect({ data, siteId: this.config.globals?.siteId });
 			this.eventManager.fire('track.product.redirect', { controller: this, redirectURL, trackEvent: data });
+			this.tracker.events.search.redirect({ data, siteId: this.config.globals?.siteId });
 		},
 	};